Move AuthSessionUpdater to datamanager auth
